Clean up AF_UNIX files on start
Summary:
If an IPC socket file exists when trying to start the socket server, assume
it's a dangling resource due to unclean shutdown and unlink() it. Obviously
this isn't 100% safe, but it's good for the 99% case.

Test Plan:
started server with and without hanging file sitting around, no
longer errors out if the file is there.

Differential Revision: https://phabricator.ericstern.com/D9